move TelcoinNode to types for now - should be in separate crate probably
