+++ crates/types/Cargo.toml
@@ -50,5 +50,13 @@ serde_yaml = { workspace = true }
 secp256k1 = { workspace = true }
 alloy = { workspace = true, features = ["genesis"] }
 
+# for node trait
+reth-db = { workspace = true }
+reth-evm = { workspace = true }
+reth-node-builder = { workspace = true }
+reth-node-ethereum = { workspace = true }
+reth-provider = { workspace = true }
+reth-trie-db = { workspace = true }
+
 [dev-dependencies]
 tempfile = { workspace = true }
