+++ crates/types/src/lib.rs
@@ -8,6 +8,7 @@ mod engine;
 mod genesis;
 mod helpers;
 mod multiaddr;
+mod node;
 mod notifier;
 mod primary;
 mod serde;
@@ -23,6 +24,7 @@ pub use engine::*;
 pub use genesis::*;
 pub use helpers::*;
 pub use multiaddr::*;
+pub use node::*;
 pub use notifier::*;
 pub use primary::*;
 pub use sync::*;
